New model.density module to hold HomologousDensity and density functions.
